muttrc: limit calls to access() to the minimum
access() was called many times on the same files due to the way the
diffent locations for muttrc were tested, so we ended up with situation
like this:

  access("/home/dkc/.muttrc-1.7.1", F_OK) = -1 ENOENT (No such file or directory)
  access("/home/dkc/.muttrc", F_OK)       = 0
  access("/home/dkc/.muttrc", F_OK)       = 0
  access("/home/dkc/.muttrc", F_OK)       = 0

  [... Tests for Muttrc ... ]

  access("/home/dkc/.muttrc", F_OK)       = 0

This commit limits the calls to access() to only once per location, and
stops looking for other locations once one has been found.

  access("/home/dkc/.muttrc-1.7+13 (f658e517960e)", F_OK) = -1 ENOENT (No such file or directory)
  access("/home/dkc/.muttrc", F_OK)       = 0

It also cleans up the case where -F is passed on the command line but
points to a non-existent file by moving the error path closer to the
actual fail condition.